show infer result & fix

+++ client/workload/workload.h
@@ -57,7 +57,7 @@ class InferWorker {
                            std::vector<size_t> concurrency);
   void FetchInferResult(Workload &workload, 
                         std::function<double_ms_t(size_t)> interval_fn, 
-                        uint32_t show_result);
+                        int64_t show_result);
   void Report(Workload &workload, int verbose, std::ostream &os);
 
 private:
@@ -125,30 +125,30 @@ class Workload {
   }
 
   // void InferMnist(size_t concurrency, std::function<double_ms_t(size_t)> interval_fn, 
-  //                 uint32_t show_result = 0);
+  //                 int64_t show_result = 0);
   // void InferMnistPoisson(size_t concurrency, double request_per_sec, 
-  //                        uint32_t show_result = 0);
+  //                        int64_t show_result = 0);
   // void InferMnistDynamic(const std::vector<double> &change_time_points, 
   //                        const std::vector<size_t> &concurrencys, 
-  //                        uint32_t show_result = 0);
+  //                        int64_t show_result = 0);
   // void InferResnet(const std::string &model, size_t concurrency, std::function<double_ms_t(size_t)> interval_fn, 
-  //                  uint32_t show_result = 0);
+  //                  int64_t show_result = 0);
   // void InferResnetPoisson(size_t concurrency, double request_per_sec, 
-  //                         uint32_t show_result = 0);
+  //                         int64_t show_result = 0);
   // void InferResnetDynamicPoisson(size_t concurrency,
   //                                const std::vector<double> &change,
   //                                const std::vector<double> &lambdas,
-  //                                uint32_t show_result = 0);
+  //                                int64_t show_result = 0);
   // void InferResnetDynamic(const std::vector<double> &change_time_points,
   //                         const std::vector<size_t> &concurrencys,
-  //                         uint32_t show_result = 0);
+  //                         int64_t show_result = 0);
 
-  void Infer(const std::string &model, size_t concurrency, std::function<double_ms_t(size_t)> interval_fn, uint32_t show_result = 0);
-  void InferPoisson(const std::string &model, size_t concurrency, double lambda, uint32_t show_result = 0);
+  void Infer(const std::string &model, size_t concurrency, std::function<double_ms_t(size_t)> interval_fn, int64_t show_result = 0);
+  void InferPoisson(const std::string &model, size_t concurrency, double lambda, int64_t show_result = 0);
   void InferDynamic(const std::string &model, 
                     const std::vector<double> &change_time_points,
                     const std::vector<size_t> &concurrencys,
-                    uint32_t show_result = 0);
+                    int64_t show_result = 0);
   void InferDynamicPoisson(const std::string &model, size_t concurrency,
                            const std::vector<double> &change_time_points,
                            const std::vector<double> &lambdas,
@@ -169,22 +169,22 @@ class Workload {
   // void Infer(const std::string &model, size_t concurrency, 
   //            std::function<void(std::vector<InferRequest>&)> set_request_fn,
   //            std::function<double_ms_t(size_t)> interval_fn,
-  //            uint32_t show_result);
+  //            int64_t show_result);
   // void InferPoisson(const std::string &model, size_t concurrency,
   //                   std::function<void(std::vector<InferRequest>&)> set_request_fn,
   //                   double request_per_sec,
-  //                   uint32_t show_result);
+  //                   int64_t show_result);
   // void InferDynamicPoisson(const std::string &model,
   //                          size_t concurrency,
   //                          const std::vector<double> &change_time_points,
   //                          const std::vector<double> &lambdas,
   //                          std::function<void(std::vector<InferRequest>&)> set_request_fn,
-  //                          uint32_t show_result);
+  //                          int64_t show_result);
   // void InferDynamic(const std::string &mode,
   //                   const std::vector<double> &change_time_points,
   //                   const std::vector<size_t> &concurrencys,
   //                   std::function<void(std::vector<InferRequest>&)> set_request_fn,
-                    // uint32_t show_result);
+                    // int64_t show_result);
 
   std::atomic<bool> running_{false};
   std::promise<void> ready_promise_;
